+++ packages/blade/src/components/Input/DropdownInputTriggers/InputChevronIcon.web.tsx
@@ -1,0 +1,28 @@
+import type { SelectChevronIconProps } from './types';
+import BaseBox from '~components/Box/BaseBox';
+import { ChevronDownIcon, ChevronUpIcon } from '~components/Icons';
+
+const InputChevronIcon = (props: SelectChevronIconProps): React.ReactElement => {
+  return (
+    <BaseBox
+      display="flex"
+      height="100%"
+      justifyContent="center"
+      alignItems="center"
+      onClick={props.onClick}
+    >
+      <ChevronDownIcon
+        display={props.isOpen ? 'none' : 'flex'}
+        color="surface.text.normal.lowContrast"
+        size="medium"
+      />
+      <ChevronUpIcon
+        display={props.isOpen ? 'flex' : 'none'}
+        color="surface.text.normal.lowContrast"
+        size="medium"
+      />
+    </BaseBox>
+  );
+};
+
+export { InputChevronIcon };