+++ packages/blade/src/components/Tag/Tag.tsx
@@ -1,72 +1,32 @@
 import React from 'react';
 import styled from 'styled-components';
+import type { TagProps } from './types';
 import { Box } from '~components/Box';
-import type { StyledPropsBlade } from '~components/Box/styledProps';
 import { getStyledProps } from '~components/Box/styledProps';
 import { IconButton } from '~components/Button/IconButton';
-import type { IconComponent } from '~components/Icons';
 import { CloseIcon } from '~components/Icons';
 import { Text } from '~components/Typography';
 import { metaAttribute, MetaConstants } from '~utils/metaAttribute';
-import type { StringChildrenType, TestID } from '~utils/types';
 import { isReactNative, makeSize } from '~utils';
 import { size as globalSizeTokens } from '~tokens/global';
 import BaseBox from '~components/Box/BaseBox';
 import type { PaddingProps } from '~components/Box/BaseBox/types/spacingTypes';
 import { useIsMobile } from '~utils/useIsMobile';
 
-type TagProps = {
-  /**
-   * Decides the size of Tag
-   *
-   * @default medium
-   */
-  size?: 'medium' | 'large';
+const FocussableTag = styled(BaseBox)<{ _isVirtuallyFocused: TagProps['_isVirtuallyFocused'] }>(
+  (props) => {
+    if (props._isVirtuallyFocused && !isReactNative()) {
+      return {
+        outline: `${makeSize(globalSizeTokens['1'])} solid ${
+          props.theme.colors.surface.background.level1.lowContrast
+        }`,
+        boxShadow: `0px 0px 0px 4px ${props.theme.colors.brand.primary[400]}`,
+      };
+    }
 
-  /**
-   * Leading icon for your Tag
-   */
-  icon?: IconComponent;
-
-  /**
-   * Callback when close icon on Tag is clicked
-   */
-  onDismiss: () => void;
-
-  /**
-   * Text that renders inside Tag
-   */
-  children: StringChildrenType;
-
-  /**
-   * Disable tag
-   */
-  isDisabled?: boolean;
-
-  /**
-   * Private property for Blade.
-   *
-   * Should not be used by consumers.
-   *
-   * Used for adding virtual focus on tag.
-   *
-   * @private
-   *
-   */
-  _isTagFocussed?: boolean;
-} & StyledPropsBlade &
-  TestID;
-
-const FocussableTag = styled(BaseBox)<{ isTagFocussed: TagProps['_isTagFocussed'] }>((props) => {
-  if (props.isTagFocussed) {
-    return {
-      outline: `1px solid ${props.theme.colors.surface.background.level1.lowContrast}`,
-      boxShadow: `0px 0px 0px 4px ${props.theme.colors.brand.primary[400]}`,
-    };
-  }
-
-  return {};
-});
+    return {};
+  },
+);
 
 /**
  * ## Tags
@@ -102,7 +62,8 @@ const Tag = ({
   children,
   isDisabled,
   testID,
-  _isTagFocussed,
+  _isVirtuallyFocused,
+  _isTagInsideInput,
   ...styledProps
 }: TagProps): React.ReactElement | null => {
   const isMobile = useIsMobile();
@@ -143,7 +104,7 @@ const Tag = ({
         backgroundColor="brand.gray.a100.lowContrast"
         borderRadius="max"
         padding={size === 'medium' ? mediumPadding : largePadding}
-        isTagFocussed={_isTagFocussed}
+        _isVirtuallyFocused={_isVirtuallyFocused}
       >
         {/* Leading Icon */}
         {Icon ? (
@@ -171,7 +132,12 @@ const Tag = ({
             icon={CloseIcon}
             accessibilityLabel={`Close ${children} tag`}
             isDisabled={isDisabled}
-            onClick={() => {
+            _tabIndex={_isTagInsideInput ? -1 : undefined}
+            onClick={(e) => {
+              // Inside tag input, we stop propagation to avoid opening dropdown on click of close on tag
+              if (_isTagInsideInput) {
+                e.stopPropagation();
+              }
               onDismiss();
             }}
           />
@@ -181,4 +147,4 @@ const Tag = ({
   );
 };
 
-export { Tag, TagProps };
+export { Tag };